refactor(variablekeywordsearch): rename GiovanniSearch to VariableKeywordSearch
