+++ cspell.json
@@ -120,6 +120,7 @@
     "radiogroup",
     "Railsbyte",
     "remixicon",
+    "renderable",
     "reregister",
     "resizer",
     "resizers",
