--- docs/pages/components/giovanni-search.md
@@ -1,22 +1,0 @@
----
-meta:
-    title: Giovanni Search
-    description:
-layout: component
----
-
-```html:preview
-<terra-giovanni-search></terra-giovanni-search>
-```
-
-## Examples
-
-### First Example
-
-TODO
-
-### Second Example
-
-TODO
-
-[component-metadata:terra-giovanni-search]
