+++ docs/pages/components/variable-keyword-search.md
@@ -1,0 +1,22 @@
+---
+meta:
+    title: Variable Keyword Search
+    description:
+layout: component
+---
+
+```html:preview
+<terra-variable-keyword-search></terra-variable-keyword-search>
+```
+
+## Examples
+
+### First Example
+
+TODO
+
+### Second Example
+
+TODO
+
+[component-metadata:terra-variable-keyword-search]
