+++ src/components/variable-keyword-search/variable-keyword-search.styles.ts
