+++ src/components/variable-keyword-search/variable-keyword-search.ts
@@ -1,0 +1,12 @@
+import TerraVariableKeywordSearch from './variable-keyword-search.component.js'
+
+export * from './variable-keyword-search.component.js'
+export default TerraVariableKeywordSearch
+
+TerraVariableKeywordSearch.define('terra-variable-keyword-search')
+
+declare global {
+    interface HTMLElementTagNameMap {
+        'terra-variable-keyword-search': TerraVariableKeywordSearch
+    }
+}
