+++ src/components/variable-keyword-search/variable-keyword-search.types.ts
