+++ src/events/events.ts
@@ -2,7 +2,7 @@ export type { TerraChangeEvent } from './terra-change.js'
 export type { TerraComboboxChangeEvent } from './terra-combobox-change.js'
 export type { TerraDateRangeChangeEvent } from './terra-date-range-change.js'
 export type { TerraErrorEvent } from './terra-error.js'
-export type { TerraGiovanniSearchChangeEvent } from './terra-giovanni-search-change.js'
+export type { TerraVariableKeywordSearchChangeEvent } from './terra-variable-keyword-search-change.js'
 export type { TerraLoadEvent } from './terra-load.js'
 export type { TerraMapEvent } from './terra-map-change.js'
 export type { TerraRemoveEvent } from './terra-remove.js'
