--- src/events/terra-giovanni-search-change.ts
@@ -1,7 +1,0 @@
-export type TerraGiovanniSearchChangeEvent = CustomEvent<string>
-
-declare global {
-    interface GlobalEventHandlersEventMap {
-        'terra-giovanni-search-change': TerraGiovanniSearchChangeEvent
-    }
-}
