+++ src/events/terra-variable-keyword-search-change.ts
@@ -1,0 +1,7 @@
+export type TerraVariableKeywordSearchChangeEvent = CustomEvent<string>
+
+declare global {
+    interface GlobalEventHandlersEventMap {
+        'terra-variable-keyword-search-change': TerraVariableKeywordSearchChangeEvent
+    }
+}
