+++ src/terra-ui-components.ts
@@ -11,7 +11,7 @@ export { default as TerraTimeSeries } from './components/time-series/time-series
 export { default as TerraVariableCombobox } from './components/variable-combobox/variable-combobox.js'
 export { default as TerraLoader } from './components/loader/loader.js'
 
-export { default as TerraGiovanniSearch } from './components/giovanni-search/giovanni-search.js';
+export { default as TerraVariableKeywordSearch } from './components/variable-keyword-search/variable-keyword-search.js'
 /* plop:component */
 
 // Utilities
